backend/requirements.txt
fastapi>=0.114.1
uvicorn>=0.30.6
requests>=2.31.0
python-dotenv>=1.0.1

